backend/requirements.txt
aiosmtpd==1.4.4.post2
aiosmtplib==2.0.1
alembic==1.9.1
anyio==3.6.2
atpublic==3.1.1
attrs==22.2.0
bcrypt==4.0.1
black==22.12.0
blinker==1.5
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.88.0
fastapi-jwt-auth==0.5.0
fastapi-mail==1.2.4
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
mypy-extensions==0.4.3
orjson==3.8.3
passlib==1.7.4
pathspec==0.10.3
platformdirs==2.6.2
psycopg2==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyJWT==1.7.1
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.6.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
